+++ src/features/App/Profiles/hooks/useProfileContainers.ts
@@ -3,8 +3,8 @@ import { useUnit } from 'effector-react';
 import { EncryptionController } from '@core/encryption/EncryptionController';
 import { PlaceholderEncryptionController } from '@core/encryption/PlaceholderEncryptionController';
 import { base64ToBytes } from '@core/encryption/utils/encoding';
+import { EncryptionAlgorithm } from '@core/features/encryption/algorithms';
 import { createEncryption } from '@core/features/encryption/createEncryption';
-import { EncryptionAlgorithm } from '@core/features/encryption/EncryptionAlgorithm';
 import { FileController } from '@core/features/files/FileController';
 import { FileControllerWithEncryption } from '@core/features/files/FileControllerWithEncryption';
 import { WorkspacesController } from '@core/features/workspaces/WorkspacesController';
